Refactor CodeEditor and split debugger functionality to it's own file
