--- src/il/co/codeguru/corewars8086/gui/nasmHandler.java
@@ -1,4 +1,0 @@
-package il.co.codeguru.corewars8086.gui;
-
-public class nasmHandler {
-}
